backend/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "synapseworth-backend"
version = "0.1.0"
description = "Backend for SynapseWorth - AI-powered skill evaluation on Hedera"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "groq==0.4.1",
    "hedera-sdk-py==0.9.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pytest==7.4.3",
    "pytest-cov==4.1.0",
    "pytest-mock==3.12.0",
    "pytest-asyncio==0.21.1",
    "httpx==0.25.2",
    "supabase==2.4.1",
    "numpy==1.26.4",
    "scipy==1.11.4",
    "pyproj==3.6.1",
    "pyModeS==2.15",
    "aiohttp==3.9.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=. --cov-report=html --cov-report=term-missing --cov-fail-under=100"
